+++ CHANGELOG.md
@@ -117,6 +117,7 @@
 - PR #4251 Add class to docs in `dask-cudf` `derived_from`
 - PR #4261 libxx Cython reorganization
 - PR #4282 Porting nvstrings conversion functions from new libcudf++ to Python/Cython
+- PR #XXXX Use libcudf++ `unary.pyx` cast instead of legacy cast
 
 ## Bug Fixes